refactor: extract contract calls chart into separate component
